Split up actions and rotations, they were getting a lot Also track potion from fflogs Tests updated too
